requirements.txt
selenium
praw
python-dotenv
fastapi
uvicorn[standard]
